[#25] fix: 상품 생성,수정,삭제 시 사용자 검증 로직 추가

+++ src/main/java/org/example/mollyapi/product/service/ProductServiceImpl.java
@@ -10,8 +10,9 @@ import org.example.mollyapi.product.entity.Product;
 import org.example.mollyapi.product.entity.ProductImage;
 import org.example.mollyapi.product.entity.ProductItem;
 import org.example.mollyapi.product.file.FileStore;
-import org.example.mollyapi.product.repository.ProductImageRepository;
 import org.example.mollyapi.product.repository.ProductRepository;
+import org.example.mollyapi.user.entity.User;
+import org.example.mollyapi.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -26,7 +27,7 @@ public class ProductServiceImpl implements ProductService {
     private final CategoryService categoryService;
     private final FileStore fileStore;
     private final ProductRepository productRepository;
-    private final ProductImageRepository productImageRepository;
+    private final UserRepository userRepository;
 
     @Override
     @Transactional
@@ -36,10 +37,18 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public Product updateProduct(Long id, ProductRegisterReqDto productRegisterReqDto) {
+    public Product updateProduct(Long userId, Long id, ProductRegisterReqDto productRegisterReqDto) {
+
+        // 유저 조회
+        User user = userRepository.findById(userId).orElseThrow(IllegalArgumentException::new);
 
         // 업데이트 할 상품 조회
-        Product product = productRepository.findById(id).orElse(null);
+        Product product = productRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
+        // 유저-상품 검증
+        if (!userId.equals(product.getUser().getUserId())) {
+            throw new IllegalArgumentException();
+        }
 
         List<ProductItem> productItems = productRegisterReqDto.items().stream().map(ProductItemReqDto::toEntity).toList();
         
@@ -57,11 +66,15 @@ public class ProductServiceImpl implements ProductService {
     @Override
     @Transactional
     public Product registerProduct(
+            Long userId,
             ProductRegisterReqDto productRegisterReqDto,
             MultipartFile thumbnailImage,
             List<MultipartFile> productImages,
             List<MultipartFile> descriptionImages
     ) {
+
+        User user = userRepository.findById(userId).orElseThrow(IllegalArgumentException::new);
+
         // 카테고리 검색
         Category category = categoryService.getCategory(productRegisterReqDto.categories());
 
@@ -73,11 +86,11 @@ public class ProductServiceImpl implements ProductService {
         // 저장된 파일로 상품 이미지 생성
         List<ProductImage> images = new ArrayList<>();
         images.add(ProductImage.createThumbnail(null, uploadThumbnail));
-        for (int i = 0; i < productRegisterReqDto.categories().size(); i++) {
+        for (int i = 0; i < uploadProductImages.size(); i++) {
             images.add(ProductImage.createProductImage(null, uploadProductImages.get(i), i+1));
         }
 
-        for (int i = 0; i < productRegisterReqDto.categories().size(); i++) {
+        for (int i = 0; i < uploadDescriptionImages.size(); i++) {
             images.add(ProductImage.createDescriptionImage(null, uploadDescriptionImages.get(i), i));
         }
 
@@ -92,6 +105,7 @@ public class ProductServiceImpl implements ProductService {
                 .description(productRegisterReqDto.description())
                 .images(images)
                 .items(items)
+                .user(user)
                 .build();
 
         return productRepository.save(newProduct);
@@ -105,7 +119,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+    public void deleteProduct(Long userId,Long id) {
+        Product product = productRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+        if (product.getUser().getUserId().equals(userId)) {
+            productRepository.deleteById(id);
+        }
     }
 }
